Add shortcut to toggle the Floating Window.
Adds a shortcut that triggers an action that triggers the service for the floating window. A transparent theme is applied to the action to prevent flickering on some devices.

This allows for macro programs to directly toggle the window instead of having to open the app or tap the tile.